pyproject.toml
[project]
name = "ghscope"
version = "0.1.0"
description = "Extensive GitHub activity and contribution reporting"
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Fuad Alizada", email = "f.alizada@alievsspace.com" }
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "jinja2>=3.1.0",
    "click>=8.1.7",
    "python-dateutil>=2.9.0",
    "weasyprint>=62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "respx>=0.21.1",
]

[project.scripts]
ghscope = "ghscope.cli:main"

[tool.mypy]
python_version = "3.12"
strict = true
pretty = true

[tool.ruff]
line-length = 100
target-version = "py312"

